+++ Natrium/logger.rb
@@ -1,14 +1,20 @@
 module Esites
   class Logger
+    @@show_timestamp = true
+
     def self._log(line, color=39, should_print=true)
-      p = "\e[90m[#{Time.now.strftime("%H:%M:%S")}]: ▸\e[0m \e[#{color}m#{line}\e[0m"
+      t = ""
+      if @@show_timestamp
+        t = "\e[90m[#{Time.now.strftime("%H:%M:%S")}]: ▸\e[0m "
+      end
+      p = "#{t}\e[#{color}m#{line}\e[0m"
       if should_print == true
         print("#{p}\n")
       end
       return p
     end
 
-    def self.error(line, should_print=true)
+    def self.error(line, should_print=true, show_time)
       return _log("[ERROR] #{line}", 31, should_print)
     end