+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog Natrium
 
+## v4.0 (23-08-2017)
+- Xcode9 and swift4 compatible
+
+## v3.0 (27-06-2017)
+- Added `idioms` to the `appicon` configuration
+- Renamed `build-config.yml` to `.natrium.yml`
+
 ## v2.3 (20-03-2017)
 - Add the [Natrium] build phase automatically